Split out workspace tests
